refactor: split pullRequests service into focused modules
Extract pullRequests.ts into ghUtils, prList, prDetail, and checkoutPRBranch — each with its own test file — so each module has a single domain responsibility. Update ProjectsController to import directly from the new modules.
